Revert "Shopping Cart: Convert lib/cart-values to TypeScript (#58609)"
This reverts commit 363af388432312376eefacf742d94247a79a1c6c.
