+++ packages/calypso-products/src/is-free-wordpress-com-domain.ts
@@ -1,3 +1,3 @@
-export function isFreeWordPressComDomain( product: { is_free?: boolean } ): boolean {
+export function isFreeWordPressComDomain( product: { is_free: boolean } ): boolean {
 	return product.is_free === true;
 }
